Extract traverse for concurrent types to its own file
